+++ lib/str-timestamp/tests/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_unit_test(LIBTEST CRITERION TARGET test_parse_timestamp)
