+++ modules/syslogformat/Makefile.am
@@ -2,8 +2,6 @@ module_LTLIBRARIES				+= modules/syslogformat/libsyslogformat.la
 modules_syslogformat_libsyslogformat_la_SOURCES	=	\
 	modules/syslogformat/syslog-format.c		\
 	modules/syslogformat/syslog-format.h		\
-	modules/syslogformat/syslog-timestamp.c		\
-	modules/syslogformat/syslog-timestamp.h		\
 	modules/syslogformat/syslog-format-plugin.c	\
 	modules/syslogformat/syslog-parser-grammar.y	\
 	modules/syslogformat/syslog-parser-parser.c	\
@@ -33,5 +31,3 @@ modules_syslogformat_libsyslogformat_la_DEPENDENCIES =	\
 modules/syslogformat modules/syslogformat/ mod-syslogformat: \
 	modules/syslogformat/libsyslogformat.la
 .PHONY: modules/syslogformat/ mod-syslogformat
-
-include modules/syslogformat/tests/Makefile.am
